+++ src/main/java/edu/ntnu/irr/bidata/GeneralClassedFrontend/ChoosAmountOfPlayers/AmountOfPlayersPage.java
@@ -1,0 +1,16 @@
+package edu.ntnu.irr.bidata.GeneralClassedFrontend.ChoosAmountOfPlayers;
+
+import javafx.scene.Scene;
+
+public class AmountOfPlayersPage extends Scene {
+    private static final AmountOfPlayersLayout layout = new AmountOfPlayersLayout();
+
+    public AmountOfPlayersPage() {
+        super(layout);  
+        if (getClass().getResource("style.css") != null) {
+            this.getStylesheets().add(getClass().getResource("style.css").toExternalForm());
+        } else {
+            System.err.println("Warning: style.css not found!");
+        }
+    }
+}
