+++ src/main/java/edu/ntnu/irr/bidata/GeneralClassedFrontend/CreatePlayer/CreatePlayerLayout.java
@@ -2,11 +2,10 @@ package edu.ntnu.irr.bidata.GeneralClassedFrontend.CreatePlayer;
 
 import javafx.scene.layout.StackPane;
 
-
 public class CreatePlayerLayout extends StackPane {
-  public CreatePlayerLayout() {
-    super();
-    this.getStyleClass().add("createPlayer-layout");
-    this.getChildren().add(new CreatePlayerCard());
-  }
+    public CreatePlayerLayout() {
+        super();
+        this.getStyleClass().add("createPlayer-layout");
+        this.getChildren().add(new CreatePlayerCard()); // Ensure CreatePlayerCard is properly initialized
+    }
 }
