+++ src/main/java/edu/ntnu/irr/bidata/GeneralClassedFrontend/CreatePlayer/CreatePlayerPage.java
@@ -1,14 +1,17 @@
 package edu.ntnu.irr.bidata.GeneralClassedFrontend.CreatePlayer;
 
-
 import javafx.scene.Scene;
 
 public class CreatePlayerPage extends Scene {
-  private static final CreatePlayerLayout layout = new CreatePlayerLayout();
+    private static final CreatePlayerLayout layout = new CreatePlayerLayout();
 
-  public CreatePlayerPage() {
-    super(layout);
-    this.getStylesheets().add(getClass().getResource("style.css").toExternalForm());
-  }
-  
+    public CreatePlayerPage() {
+        super(layout);
+
+        if (getClass().getResource("/style.css") != null) {
+            this.getStylesheets().add(getClass().getResource("style.css").toExternalForm());
+        } else {
+            System.err.println("Warning: style.css not found!");
+        }
+    }
 }
