+++ src/main/java/edu/ntnu/irr/bidata/GeneralClassedFrontend/UI.java
@@ -1,11 +1,11 @@
 package edu.ntnu.irr.bidata.GeneralClassedFrontend;
 
-import edu.ntnu.irr.bidata.GeneralClassedFrontend.ChoosAmoutOfPlayers.AmoutOfPlayersPage;
+import edu.ntnu.irr.bidata.GeneralClassedFrontend.ChoosAmountOfPlayers.AmountOfPlayersPage;
 import edu.ntnu.irr.bidata.GeneralClassedFrontend.CreatePlayer.CreatePlayerPage;
 
 public class UI {
-  private static final CreatePlayerPage createPlayer = new CreatePlayerPage();
-  private static final AmoutOfPlayersPage AmoutOfPlayers = new AmoutOfPlayersPage();
+  //private static final CreatePlayerPage createPlayer = new CreatePlayerPage();
+  private static final AmountOfPlayersPage AmountOfPlayers = new AmountOfPlayersPage();
 
   public UI() {
 
@@ -19,24 +19,22 @@ public class UI {
   public static void toAmountOfPlayersPage() {
     System.out.println("Switching to AmountOfPlayersPage...");
     try {
-        MyWindow.getPrimaryStage().hide();
-        MyWindow.getPrimaryStage().setMaximized(false);
-        MyWindow.getPrimaryStage().setScene(AmoutOfPlayers);
-        MyWindow.getPrimaryStage().setMaximized(true);
-        MyWindow.getPrimaryStage().show();
-    } catch (Exception e) {
-        e.printStackTrace();
-    }
+      MyWindow.getPrimaryStage().hide();
+      MyWindow.getPrimaryStage().setMaximized(false);
+      MyWindow.getPrimaryStage().setScene(AmountOfPlayers);
+      MyWindow.getPrimaryStage().setMaximized(true);
+      MyWindow.getPrimaryStage().show();
+  } catch (Exception e) {
+      e.printStackTrace();
+  }        
 }
 
   private static void toCreatePlayerPage() {
     MyWindow.getPrimaryStage().hide();
     MyWindow.getPrimaryStage().setMaximized(false);
-    MyWindow.getPrimaryStage().setScene(createPlayer);
+    //MyWindow.getPrimaryStage().setScene(createPlayer);
     MyWindow.getPrimaryStage().setMaximized(true);
     MyWindow.getPrimaryStage().show();
   }
-
-
   
 }
