+++ src/test/java/edu/ntnu/irr/bidata/DieTest.java
@@ -12,17 +12,17 @@
 
 // public class DieTest {
 
-  @Test
-  @DisplayName("Max and min values are set correctly")
-  void testMaxAndMinValuesOfRoll() {
-    Die myDie = new Die(6);
-    boolean areAllRollsInRange = true;
-    for (int i = 0; i < 1000; i++) {
-      int roll = myDie.roll();
-      if (roll > 6 || roll < 1) {
-        areAllRollsInRange = false;
-      }
-    }
-    assertTrue(areAllRollsInRange, "All rolls are in a valid range between 1 and 6");
-  }
-}
+//   @Test
+//   @DisplayName("Max and min values are set correctly")
+//   void testMaxAndMinValuesOfRoll() {
+//     Die myDie = new Die(6);
+//     boolean areAllRollsInRange = true;
+//     for (int i = 0; i < 1000; i++) {
+//       int roll = myDie.roll();
+//       if (roll > 6 || roll < 1) {
+//         areAllRollsInRange = false;
+//       }
+//     }
+//     assertTrue(areAllRollsInRange, "All rolls are in a valid range between 1 and 6");
+//   }
+// }
